requirements.txt
amqp==5.0.5
asgiref==3.3.1
backcall==0.2.0
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Django==3.1.7
django-extensions==3.1.1
django-shell-plus==1.1.7
dnspython==2.1.0
filelock==3.0.12
flake8==3.8.4
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==3.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
mccabe==0.6.1
numpy==1.20.1
oauth2client==4.1.3
packaging==20.9
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.17
protobuf==3.15.6
ptyprocess==0.7.0
py3-validate-email==1.0.0
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
